requirements.txt
amqp==5.2.0
asgiref==3.8.1
awscli==1.34.11
billiard==4.2.0
boto3==1.35.11
botocore==1.35.11
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.16
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis-cache==3.0.1
django-ses==4.1.1
django-test-without-migrations==0.6
djangorestframework==3.15.2
docutils==0.16
drf-yasg==1.21.7
grpcio==1.66.1
grpcio-tools==1.66.1
idna==3.8
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.0
packaging==24.1
prompt_toolkit==3.0.47
protobuf==5.28.0
psycopg==3.2.1
pyasn1==0.6.0
pylibmc==1.6.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==3.5.3
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
setuptools==74.1.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uWSGI==2.0.26
vine==5.1.0
wcwidth==0.2.13
wheel==0.44.0
